+++ src/scicloj/clay/v2/notebook.clj
@@ -1,17 +1,16 @@
 (ns scicloj.clay.v2.notebook
-  (:require [clojure.string :as string]
-            [scicloj.clay.v2.item :as item]
+  (:require [clojure.string :as str]
+            [clojure.pprint :as pp]
             [scicloj.clay.v2.util.path :as path]
             [scicloj.clay.v2.item :as item]
             [scicloj.clay.v2.prepare :as prepare]
             [scicloj.clay.v2.read :as read]
-            [scicloj.clay.v2.config :as config]
-            [scicloj.clay.v2.util.merge :as merge]
             [scicloj.kindly.v4.api :as kindly]
             [scicloj.kindly.v4.kind :as kind]
-            [scicloj.kindly-advice.v1.api :as kindly-advice]
-            [clojure.string :as str]
-            [clojure.pprint :as pp]))
+            [scicloj.kindly-advice.v1.api :as kindly-advice])
+  (:import (java.io StringWriter)))
+
+(set! *warn-on-reflection* true)
 
 (defn deref-if-needed [v]
   (if (delay? v)
@@ -29,7 +28,7 @@
                          remote-repo]}]
   (let [relative-file-path (path/path-relative-to-repo full-source-path)]
     (item/info-line {:path relative-file-path
-                     :url (some-> remote-repo (path/file-git-url relative-file-path))})))
+                     :url  (some-> remote-repo (path/file-git-url relative-file-path))})))
 
 (defn narrowed? [code]
   (some-> code
@@ -39,34 +38,75 @@
   (some-> code
           (str/includes? ",,,")))
 
-(defn complete [{:as note
-                 :keys [comment? code form value]}]
-  (-> (if (or value comment?)
-        note
-        (assoc note
-               :value (cond code (-> code
-                                     read-string
-                                     eval
-                                     deref-if-needed)
-                            form (-> form
-                                     eval
-                                     deref-if-needed))))
-      (cond-> (not comment?)
-        kindly-advice/advise)))
+(defn ns-form? [form]
+  (and (sequential? form)
+       (-> form first (= 'ns))))
+
+(defn str-and-reset! [w]
+  (when (instance? StringWriter *out*)
+    (locking w
+      (let [s (str w)]
+        (.setLength (.getBuffer ^StringWriter w) 0)
+        s))))
+
+(defn read-eval-capture
+  "Captures stdout and stderr while evaluating a note"
+  [{:as   note
+    :keys [code form]}]
+  (let [out (StringWriter.)
+        err (StringWriter.)
+        note (try
+               (let [x (binding [*out* out
+                                 *err* err]
+                         (cond code (-> code
+                                        read-string
+                                        eval
+                                        deref-if-needed)
+                               form (-> form
+                                        eval
+                                        deref-if-needed)))]
+                 (assoc note :value x))
+               (catch Throwable ex
+                 (assoc note :exception ex)))
+        out-str (str out)
+        err-str (str err)
+        ;; A notebook may have also printed from a thread,
+        ;; *out* and *err* are replaced with StringWriters in with-out-err-capture
+        global-out (str-and-reset! *out*)
+        global-err (str-and-reset! *err*)
+        ;; Don't show output from requiring other namespaces
+        show (not (ns-form? form))]
+    (if show
+      (cond-> note
+        (seq out-str) (assoc :out out-str)
+        (seq err-str) (assoc :err err-str)
+        (seq global-out) (assoc :global-out global-out)
+        (seq global-err) (assoc :global-err global-err))
+      note)))
+
+(defn complete [{:as   note
+                 :keys [comment?]}]
+  (let [completed (cond-> note
+                    (not (or comment? (contains? note :value)))
+                    (read-eval-capture))]
+    (cond-> completed
+      (and (not comment?) (contains? completed :value))
+      (kindly-advice/advise))))
 
 (defn comment->item [comment]
   (-> comment
-      (string/split #"\n")
+      (str/split #"\n")
       (->> (map #(-> %
-                     (string/replace
-                      #"^;+\s?" "")
-                     (string/replace
-                      #"^#" "\n#")))
-           (string/join "\n"))
+                     (str/replace
+                       #"^;+\s?" "")
+                     (str/replace
+                       #"^#" "\n#")))
+           (str/join "\n"))
       item/md))
 
-(defn hide-code? [{:as note :keys [code form value kind]} {:as opts :keys [hide-code]}]
+(defn hide-code? [{:as note :keys [code form value kind narrowed]} {:as opts :keys [hide-code]}]
   (or hide-code
+      narrowed
       (-> form meta :kindly/hide-code)
       (-> form meta :kindly/hide-code?)                     ; legacy convention
       (-> value meta :kindly/hide-code)
@@ -104,22 +144,30 @@
       :deps   (set (mapcat :deps value-items))}]))
 
 (defn note-to-items [{:as   note
-                      :keys [comment? code
+                      :keys [comment?
+                             code
+                             exception
+                             err
+                             out
+                             global-err
+                             global-out
                              kindly/options]}
-                     {:as   opts}]
+                     {:as opts}]
   (if (and comment? code)
     [(comment->item code)]
     (let [code-item (when-not (hide-code? note opts)
                       (item/source-clojure code))
-          value-items (when-not (hide-value? note opts)
-                        (-> note
-                            (select-keys [:value :code :form
-                                          :base-target-path
-                                          :full-target-path
-                                          :kindly/options
-                                          :format])
-                            (update :value deref-if-needed)
-                            prepare/prepare-or-pprint))]
+          value-items (cond-> []
+                              err (conj (item/print-output "ERR" err))
+                              out (conj (item/print-output "OUT" out))
+                              global-err (conj (item/print-output "THREAD ERR" global-err))
+                              global-out (conj (item/print-output "THREAD OUT" global-out))
+                              exception (conj (item/print-throwable exception))
+                              (and (contains? note :value)
+                                   (not (hide-value? note opts)))
+                              (into (-> note
+                                        (update :value deref-if-needed)
+                                        (prepare/prepare-or-pprint))))]
       (cond (and (not code-item) (empty? value-items))
             []
 
@@ -129,7 +177,8 @@
             (empty? value-items)
             [code-item]
 
-            (= :horizontal (or (:code-and-value opts) (:code-and-value options)))
+            (= :horizontal (or (:code-and-value opts)
+                               (:code-and-value options)))
             (side-by-side-items opts code-item value-items)
 
             :else
@@ -139,8 +188,7 @@
   (if hide-info-line
     items
     (let [il (info-line spec)]
-      (concat items
-              [item/separator il]))))
+      (into items [item/separator il]))))
 
 (defn ->var-name [i line-number]
   (symbol (str "var" i
@@ -181,10 +229,6 @@
                                         var-name)
                                   args)))))))
 
-(defn ns-form? [form]
-  (and (sequential? form)
-       (-> form first (= 'ns))))
-
 (defn test-ns-form [[_ ns-symbol & rest-ns-form]]
   (concat (list 'ns
                 (-> ns-symbol
@@ -215,25 +259,90 @@
          path
          (fn [{:keys [code]}]
            (let [new-code (slurp path)]
-             {:code new-code
+             {:code                 new-code
               :first-line-of-change (first-line-of-change
-                                     code new-code)})))
+                                      code new-code)})))
   (@*path->last path))
 
+(defmacro with-out-err-captured
+  "Evaluates and computes the items for a notebook of notes"
+  [& body]
+  ;; For a notebook, we capture output globally, and per note.
+  ;; see read-eval-capture for why this is relevant.
+  `(let [out# (StringWriter.)
+         err# (StringWriter.)]
+     ;; Threads may inherit only the root binding
+     (with-redefs [*out* out#
+                   *err* err#]
+       ;; Futures will inherit the current binding,
+       ;; which was not affected by altering the root.
+       (binding [*out* out#
+                 *err* err#]
+         ~@body))))
+
+(defn complete-notes
+  "Evaluates and computes the items for a notebook of notes"
+  [relevant-notes some-narrowed options]
+  (reduce (fn [{:as aggregation :keys [i
+                                       items
+                                       test-forms
+                                       last-nontest-varname]}
+               note]
+            (let [{:as complete-note :keys [form region narrowed exception comment?]} (complete note)
+                  test-note (test-last? complete-note)
+                  new-items (when (or (not some-narrowed)
+                                      narrowed)
+                              (when-not test-note
+                                (-> complete-note
+                                    (kindly/deep-merge
+                                      (-> options
+                                          (select-keys [:base-target-path
+                                                        :full-target-path
+                                                        :kindly/options
+                                                        :format])))
+                                    (note-to-items options))))
+                  line-number (first region)
+                  varname (->var-name i line-number)
+                  test-form (cond
+                              ;; a deftest form
+                              test-note (deftest-form
+                                          (->test-name i line-number)
+                                          last-nontest-varname
+                                          form)
+                              ;; the test ns form
+                              (ns-form? form) (test-ns-form form)
+                              ;; a comment
+                              comment? nil
+                              ;; the regular case, just a def
+                              :else (def-form varname form))
+                  step {:i                    (inc i)
+                        :items                (into items (remove nil?) new-items)
+                        :test-forms           (if test-form
+                                                (conj test-forms test-form)
+                                                test-forms)
+                        :last-nontest-varname (if (or comment? test-note)
+                                                last-nontest-varname
+                                                varname)
+                        :exception exception}]
+              (if (and exception (not (:exception-continue options)))
+                (reduced step)
+                step)))
+          ;; initial value
+          {:i              0
+           :items          []
+           :test-forms     []
+           :last-nontest-i nil}
+          ;; sequence
+          relevant-notes))
+
 (defn items-and-test-forms
-  ([{:as options
+  ([{:as   options
      :keys [full-source-path
-            hide-info-line
-            hide-code hide-nils hide-vars
-            title toc?
-            base-target-path
-            full-target-path
             single-form
             single-value
-            format
             smart-sync
             pprint-margin]
-     :or {pprint-margin pp/*print-right-margin*}}]
+     :or   {pprint-margin pp/*print-right-margin*}}]
    (binding [*ns* *ns*
              *warn-on-reflection* *warn-on-reflection*
              *unchecked-math* *unchecked-math*
@@ -248,7 +357,7 @@
                                                 [{:form (read/read-ns-form code)}])
                                               {:form single-form})
                             :else (read/->notes code))
-                      (map-indexed (fn [i {:as note
+                      (map-indexed (fn [i {:as   note
                                            :keys [code]}]
                                      (merge note
                                             {:i i}
@@ -279,71 +388,15 @@
                                            (or (ns-form? form)
                                                (>= i first-narrowed-index)
                                                (-> region
-                                                   (nth 2) ;last region line
+                                                   (nth 2)  ;last region line
                                                    (> first-line-of-change))))))
                             ;;
                             :else
                             notes)]
-       (-> (->> relevant-notes
-                (reduce (fn [{:as aggregation :keys [i
-                                                     items
-                                                     test-forms
-                                                     last-nontest-varname]}
-                             note]
-                          (let [{:as complete-note :keys [form kind region narrowed]} (complete note)
-                                test-note (test-last? complete-note)
-                                comment (:comment? complete-note)
-                                new-items (when (or (not some-narrowed)
-                                                    narrowed)
-                                            (when-not test-note
-                                              (-> complete-note
-                                                  (merge/deep-merge
-                                                   (-> options
-                                                       (select-keys [:base-target-path
-                                                                     :full-target-path
-                                                                     :kindly/options
-                                                                     :format])))
-                                                  (note-to-items
-                                                   (merge options
-                                                          (when narrowed
-                                                            {:hide-code true}))))))
-                                line-number (first region)
-                                varname (->var-name i line-number)
-                                test-form (cond
-                                            ;; a deftest form
-                                            test-note (deftest-form
-                                                        (->test-name i line-number)
-                                                        last-nontest-varname
-                                                        form)
-                                            ;; the test ns form
-                                            (ns-form? form) (test-ns-form form)
-                                            ;; a comment
-                                            comment nil
-                                            ;; the regular case, just a def
-                                            :else (def-form
-                                                    varname
-                                                    form))]
-                            {:i              (inc i)
-                             :items          (concat items new-items)
-                             :test-forms     (if test-form
-                                               (conj test-forms test-form)
-                                               test-forms)
-                             :last-nontest-varname (if (or comment
-                                                           test-note)
-                                                     last-nontest-varname
-                                                     varname)}))
-                        ;; initial value
-                        {:i              0
-                         :items          []
-                         :test-forms     []
-                         :last-nontest-i nil}))
-           (update :items
-                   ;; final processing of items
-                   (fn [items]
-                     (-> items
-                         (->> (remove nil?))
-                         (add-info-line options)
-                         doall)))
+       (-> relevant-notes
+           (complete-notes some-narrowed options)
+           (with-out-err-captured)
+           (update :items add-info-line options)
            (update :test-forms
                    ;; Leave the test-form only when
                    ;; at least one of them is a `deftest`.
@@ -361,4 +414,4 @@
 
   (-> "notebooks/scratch.clj"
       (notebook-items {:full-target-path "docs/scratch.html"
-                       :single-form '(+ 1 2)})))
+                       :single-form      '(+ 1 2)})))